# api/openapi.yaml
openapi: 3.1.0
info:
  title: Moirai Public API
  version: 0.1.0
  description: |
    HTTP gateway for the Moirai control plane. The API uses the `loop_session`
    cookie created by the login endpoint. Mutating authenticated requests also
    require the `X-CSRF-Token` header to equal the `loop_csrf` cookie value.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/workflows:
    get:
      summary: List workflows
      description: |
        Every row carries the same fields as `GET /api/v1/workflows/{workflow_id}`.
        The console renders issue titles, pull requests and attempt budgets in the
        list, the overview and the phase threads; a summary shape would cost one
        detail request per workflow.
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Workflows
          content:
            application/json:
              schema:
                type: object
                required: [workflows]
                properties:
                  workflows:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorkflowDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/workflows/{workflow_id}:
    get:
      summary: Get workflow details
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/WorkflowID'
      responses:
        '200': { description: Workflow details, content: { application/json: { schema: { $ref: '#/components/schemas/WorkflowDetail' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/workflows/{workflow_id}/events:
    get:
      summary: List workflow events newest first
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/WorkflowID'
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 100 }
      responses:
        '200': { description: Workflow events, content: { application/json: { schema: { $ref: '#/components/schemas/WorkflowEventsPage' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/workflows/{workflow_id}/decision:
    post:
      summary: Submit a required human decision
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - $ref: '#/components/parameters/WorkflowID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision:
                  type: string
                  enum: [approved, changes_requested]
                comment:
                  type: string
      responses:
        '200':
          description: Workflow after the decision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/workflows/{workflow_id}/retry:
    post:
      summary: Retry a terminal workflow
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - $ref: '#/components/parameters/WorkflowID'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkflowControlRequest' }
      responses:
        '200': { description: Workflow resumed, content: { application/json: { schema: { $ref: '#/components/schemas/Workflow' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/v1/workflows/{workflow_id}/cancel:
    post:
      summary: Cancel an in-flight workflow
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - $ref: '#/components/parameters/WorkflowID'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkflowControlRequest' }
      responses:
        '200': { description: Workflow cancelled, content: { application/json: { schema: { $ref: '#/components/schemas/Workflow' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/v1/workflows/{workflow_id}/block:
    post:
      summary: Block an in-flight workflow
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - $ref: '#/components/parameters/WorkflowID'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkflowControlRequest' }
      responses:
        '200': { description: Workflow blocked, content: { application/json: { schema: { $ref: '#/components/schemas/Workflow' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: loop_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  parameters:
...
    WorkflowID:
      name: workflow_id
      in: path
      required: true
      schema: { type: string }
...
  responses:
    BadRequest:
      description: Malformed request
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Authentication is required or invalid
    Forbidden:
      description: A CSRF token is required for this mutation
    NotFound:
      description: Requested resource was not found
    ValidationError:
      description: The control plane rejected valid JSON input
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    WorkflowControlRequest:
      type: object
      properties:
        reason: { type: string, maxLength: 1024 }
    Workflow:
      type: object
      required: [id, projectId, status, phase]
      properties:
        id: { type: string }
        projectId: { type: string }
        status: { type: string }
        phase: { type: string }
    WorkflowDetail:
      allOf:
        - $ref: '#/components/schemas/Workflow'
        - type: object
          required: [issueExternalId, issueTitle, branchName, planningAttempts, implementationAttempts, pipelineRepairAttempts, ciRepairAttempts, reviewCycles, totalAgentExecutions, planSummary, createdAt, updatedAt]
          properties:
            issueExternalId: { type: string }
            issueTitle: { type: string }
            branchName: { type: string }
            pullRequestExternalId: { type: string }
            pullRequestUrl: { type: string }
            pullRequestState: { type: string }
            blockingReason: { type: string }
            planningAttempts: { type: integer, minimum: 0 }
            implementationAttempts: { type: integer, minimum: 0 }
            pipelineRepairAttempts: { type: integer, minimum: 0 }
            ciRepairAttempts: { type: integer, minimum: 0 }
            reviewCycles: { type: integer, minimum: 0 }
            totalAgentExecutions:
              type: integer
              minimum: 0
              description: >-
                Agent executions spent across every role. The cap it is measured
                against lives in the orchestrator's RetryBudget.
            planSummary:
              type: string
              description: >-
                The planning phase's own account of the work (#351): the
                planner execution's summary, or empty when the project never
                opted into requirePlanning or the run has not reached that
                phase yet.
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
    WorkflowEvent:
      type: object
      required: [id, type, createdAt, payload]
      properties:
        id: { type: string }
        type: { type: string }
        createdAt: { type: string, format: date-time }
        payload: {}
    WorkflowEventsPage:
      type: object
      required: [events]
      properties:
        events:
          type: array
          items: { $ref: '#/components/schemas/WorkflowEvent' }
        nextCursor: { type: string }
...